src: refactor filter.rs into configuration.rs and respective contexts
